--- a/examples/gpt-2/CMakeLists.txt
+++ b/examples/gpt-2/CMakeLists.txt
@@ -46,3 +46,7 @@ endif()
 if (GGML_METAL)
     add_compile_definitions(GGML_USE_METAL)
 endif()
+
+if (GGML_RPC)
+    add_compile_definitions(GGML_USE_RPC)
+endif()