--- a/examples/rpc/CMakeLists.txt
+++ b/examples/rpc/CMakeLists.txt
@@ -1,0 +1,5 @@
+add_executable(client client.cpp)
+target_link_libraries(client PRIVATE ggml)
+
+add_executable(server server.cpp)
+target_link_libraries(server PRIVATE ggml)