We have a use case where we want to build and run ggml programs on low end machines (without GPUs) and leverage the computational resources of some high end machines (with GPUs) over the network. In this PR I am trying to prototype an RPC backend which proxies all operations to another host. On the remote host, the RPC backend simply delegates to one of the existing backends (CUDA, Metal, etc.):

```mermaid
flowchart TB
    rpcc---|gRPC|rpcs
    subgraph hosta[Host B]
    rpcs[RPC Server]---backend["Backend (CUDA,Metal,etc.)"]
    end
    subgraph hostb[Host A]
    ggml[ggml program]---rpcc[RPC Client]
    end
```

I am using gRPC for the remote calls, you can find the interface definition in `ggml-rpc.proto`. I have a simple program (client.cpp) which creates some tensors and successfully stores and retrieves data into them using the RPC backend.

You can give it a try with the following steps:
1. Install gRPC by following [this guide](https://grpc.io/docs/languages/cpp/quickstart/)
2. Build ggml with:
```
cmake -DGGML_RPC=ON -DCMAKE_PREFIX_PATH=$MY_INSTALL_DIR -DGGML_CUBLAS=ON -DCMAKE_BUILD_TYPE=Debug ..
make
```
With this configuration the RPC backend will delegate to the CUDA backend
3. Start the RPC server:
```
bin/server 50051
```
4. Run the sample client:
```
bin/client localhost:50051
```

I am currently looking for some guidance on how to implement `graph_compute` with this approach. Any help is appreciated.