@slaren 2024-03-07
Very cool! To implement `graph_compute`, you would need to re-create the same graph in the server. Most of the attributes of the tensors you should be able to copy them as they are in the client, but you would need to map the buffer pointers to the local copy of the buffer, as well as the tensor pointers. Roughly, something like this should work:

```python
tensor map_tensor(t):
    if t in tensor_map:
        return tensor_map[t]
    tensor new_t = t
    new_t->view_src = map_tensor(t->view_src)
    new_t->buffer = map_buffer(t->buffer)
    for i in range(GGML_MAX_SRC):
        new_t->src[i] = map_tensor(t->src[i])
    tensor_map[t] = new_t
    return new_t
}

for i in range(remote_graph->n_nodes):
    local_graph->nodes[i] = map_tensor(remote_graph->nodes[i])
```

For the CUDA backend, it is important to also wrap the buffer `init_tensor` function, and update the local tensor with the values changed by the backend (should only be the `extra` and `backend` attributes).