@rgerganov 2024-03-08
Thanks for the hints. I have added a `Tensor` protobuf structure and the client-side implementation of `graph_compute` is serializing the graph into an array of `Tensor` structures. I am using the local address of the tensor as unique identifier, so I can reconstruct the same relations on the server. As for the tensor buffer, I am keeping the remote address of the buffer in the local buffer context, so I take it from there. The `data` pointer is already a remote address, so I copy it as is.