@slaren 2024-03-11
What backend are you wrapping in the server? The CPU backend should be the simplest to make work.