@slaren 2024-03-11
The issue may be that `tensor->nb` is not serialized. For non-contiguous views this is important.