@slaren 2024-03-11
I get this error when building (followed by a million more):
```
/home/diego/usr/include/absl/strings/string_view.h:53:26: error: ‘string_view’ in namespace ‘std’ does not name a type
   53 | using string_view = std::string_view;
      |                          ^~~~~~~~~~~
/home/diego/usr/include/absl/strings/string_view.h:53:21: note: ‘std::string_view’ is only available from C++17 onwards
   53 | using string_view = std::string_view;
      |                     ^~~
```

It seems that gRPC requires building with C++17. I am missing something?