@rgerganov 2024-03-11
> The issue may be that `tensor->nb` is not serialized. For non-contiguous views this is important.

That was exactly the problem, gpt-2 works now!