@slaren 2024-03-11
> Build gRPC by adding `-DCMAKE_CXX_STANDARD=14` to cmake

Does it work with 11 too? Usually ggml targets C++11.