@rgerganov 2024-03-11
gRPC doesn't build with `-DCMAKE_CXX_STANDARD=11` but it works fine with ggml when built with `-DCMAKE_CXX_STANDARD=14`