@rgerganov 2024-03-11
Yes, these exact 3 functions are called many many times. `GetAllocSize` and `BufferGetBase` are easy to handle and `InitTensor` will require some tricks as you already suggested. When I cache the result for the first two and skip the third one (only with CPU backend), I get the following:
```
main:     load time =  1615.98 ms
main:   sample time =    17.24 ms
main:  predict time =  2319.07 ms / 33.13 ms per token
main:    total time =  3957.53 ms
```

The same prompt with CPU backend gives:
```
main:     load time =   140.88 ms
main:   sample time =     6.73 ms
main:  predict time =   536.77 ms / 7.67 ms per token
main:    total time =   685.79 ms
```