@slaren 2024-03-11
Actually `GetAllocSize` cannot be buffered, since ggml-alloc needs to know the value immediately. Not sure how to address that. A simply hack for now could be to return `ggml_nbytes`, which will be correct most of the time.