@ggerganov 2024-03-11
Metal backend also works:

<img width="1102" alt="image" src="https://github.com/ggerganov/ggml/assets/1991296/cce2eead-c413-46d5-9741-965f7714a16b">

This is the performance without gRPC:

```
main:     load time =  4307.36 ms
main:   sample time =    16.92 ms
main:  predict time =  2078.37 ms / 14.24 ms per token
main:    total time =  6403.82 ms
```